build: split main.go file into platform specific
The windows registry package doesn't build on Linux. It errors out with:
 main.go:16:2: build constraints exclude all Go files in /home/user/go/pkg/mod/golang.org/x/sys@v0.0.0-20201218084310-7d0127a74742/windows/registry

Splitting the main.go file in platform specific modules seems to be the
more elegant approach that conditional imports, which aren't possible (from
what I've read on the web).

Took inspiration for splitting these up from here:
   https://dave.cheney.net/2013/10/12/how-to-use-conditional-compilation-with-the-go-build-tool

Updated the 'build.sh' script to try some auto-detection of the build
platform.
